pyproject.toml
[project]
name = "iconfig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "heinerlehr", email = "heiner.lehr@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
iconfig = "iconfig:main"

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.1",
    "ruff>=0.14.4",
]
